backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
pillow
reportlab
pdfplumber
pytesseract
requests
aiofiles
pydantic-settings
python-dotenv
svgwrite
matplotlib
